Rename placeholders to system components
We might simplify this further at a later date, but we cannot change the depth of our directories yet or our docs will break
