+++ packages/cascara/package.json
@@ -41,6 +41,7 @@
     "color": "3.1.3",
     "color-string": "1.5.5",
     "memoize-one": "5.2.1",
+    "react-error-boundary": "3.1.3",
     "react-fast-compare": "3.2.0",
     "react-hook-form": "6.8.4",
     "react-textarea-autosize": "8.3.3",
