+++ packages/cascara/src/index.js
@@ -2,8 +2,8 @@
 // If we create any additional wildcard exports for a directory, make sure
 // we put a note at the top of the file to warn ourselves that everything
 // in that file is going to get exported.
-export * from './placeholders';
 export * from './structures';
+export * from './system-components';
 export * from './ui';
 
 // This is temporary while we figure out where this component should live.
