--- packages/cascara/src/placeholders/index.js
@@ -1,5 +1,0 @@
-// HEADS UP!
-// Everything exported here is being exported from Cascara
-
-export { default as AreaPlaceholder } from './AreaPlaceholder';
-export { default as JsonPlaceholder } from './JsonPlaceholder';
