+++ packages/cascara/src/system-components/AreaPlaceholder/AreaPlaceholder.doc.mdx
