+++ packages/cascara/src/system-components/AreaPlaceholder/AreaPlaceholder.fixture.js
