+++ packages/cascara/src/system-components/AreaPlaceholder/AreaPlaceholder.js
