+++ packages/cascara/src/system-components/AreaPlaceholder/AreaPlaceholder.module.scss
