+++ packages/cascara/src/system-components/AreaPlaceholder/index.js
