+++ packages/cascara/src/system-components/Boundaries/Boundaries.js
@@ -1,0 +1,18 @@
+import React from 'react';
+import pt from 'prop-types';
+import { ErrorBoundary } from 'react-error-boundary';
+import ErrorFallback from './ErrorFallback';
+
+const propTypes = {
+  children: pt.oneOfType([pt.arrayOf(pt.node), pt.node]),
+};
+
+const Boundaries = ({ children }) => {
+  return (
+    <ErrorBoundary FallbackComponent={ErrorFallback}>{children}</ErrorBoundary>
+  );
+};
+
+Boundaries.propTypes = propTypes;
+
+export default Boundaries;
