+++ packages/cascara/src/system-components/Boundaries/index.js
@@ -1,0 +1,1 @@
+export { default } from './Boundaries';
