+++ packages/cascara/src/system-components/index.js
@@ -1,0 +1,7 @@
+// HEADS UP!
+// Everything exported here is being exported from Cascara.
+// Let's keep these sorted alphabetically based on their source directory.
+
+export { default as AreaPlaceholder } from './AreaPlaceholder';
+export { default as Boundaries } from './Boundaries';
+export { default as JsonPlaceholder } from './JsonPlaceholder';
