+++ packages/cascara/src/ui/Pagination/Pagination.fixture.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import Pagination, { usePaginationState } from './';
-import JsonPlaceholder from '../../placeholders/JsonPlaceholder';
+import { JsonPlaceholder } from '../../system-components';
 
 const Loading = (fixtureProps) => (
   <>
